IRIS Flower Classification ML Project
For this model, the accuracy on the test set is 0.97, which means the model made the right prediction for 97% of the irises in the given dataset. We can expect the model to be correct 97% of the time for predicting the species of new irises.
For a hobby botanist application, this is a high level of accuracy and it means that our model may be trustworthy enough to use.